+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/rules/PropertyNamingRule.kt
@@ -81,7 +81,13 @@ public class PropertyNamingRule :
     ) {
         identifier
             .text
-            .takeUnless { it.matches(SCREAMING_SNAKE_CASE_REGEXP) }
+            .takeUnless {
+                // Allow
+                // object Foo {
+                //     private const val serialVersionUID: Long = 123
+                // }
+                it == SERIAL_VERSION_UID_PROPERTY_NAME
+            }?.takeUnless { it.matches(SCREAMING_SNAKE_CASE_REGEXP) }
             ?.let {
                 emit(
                     identifier.startOffset,
@@ -143,6 +149,7 @@ public class PropertyNamingRule :
         val LOWER_CAMEL_CASE_REGEXP = "[a-z][a-zA-Z0-9]*".regExIgnoringDiacriticsAndStrokesOnLetters()
         val SCREAMING_SNAKE_CASE_REGEXP = "[A-Z][_A-Z0-9]*".regExIgnoringDiacriticsAndStrokesOnLetters()
         val BACKING_PROPERTY_LOWER_CAMEL_CASE_REGEXP = "_[a-z][a-zA-Z0-9]*".regExIgnoringDiacriticsAndStrokesOnLetters()
+        const val SERIAL_VERSION_UID_PROPERTY_NAME = "serialVersionUID"
     }
 }